[scripts] Update plfix script
The lua subscripts have been updated to work with the new playcount file
structure, and now they both respect the -p pretend flag. Also, they
remove invalid file paths and are able to report that in the log file as
well.
Basically, everything works better and more consistently.

+++ scripts/.scripts/plfix_playcount.lua
@@ -1,33 +1,47 @@
 #!/usr/bin/env lua
--- This script gets rid of duplicate lines in playcount.tsv by adding any
--- duplicate entry to the first entry. It can be used on its own and is
--- automatically invoked by the plfix script.
+-- This script gets rid of duplicate lines in a playcount file by adding
+-- any duplicate entry to the first entry. It is automatically invoked by the
+-- plfix script.
 
 musicdir = os.getenv('HOME') .. '/Music/'
-fpath = musicdir .. 'playcount.tsv'
+fpath = arg[1]
+pretend = (arg[2] == '1')
 hashset = {}
 hashset_order = {}
 dupes = 0
+header = false
 
 for line in io.lines(fpath) do
 	local f = line:gmatch('[^\t]*')
 	local num, song = tonumber(f()), f()
-	if not hashset[song] then
-		hashset_order[#hashset_order + 1] = song
-    else
-        dupes = dupes + 1
-    end
-	hashset[song] = (hashset[song] or 0) + num
+
+	f = io.open(musicdir..song, 'r')
+	if f == nil then
+		if not header then
+			print(fpath:gsub(musicdir, '')..'')
+			header = true
+		end
+		print(num, song)
+	else
+		f:close()
+		if not hashset[song] then
+			hashset_order[#hashset_order + 1] = song
+		else
+			dupes = dupes + 1
+		end
+		hashset[song] = (hashset[song] or 0) + num
+	end
 end
 
-fout = io.open(fpath, 'w')
-for _, v in ipairs(hashset_order) do
-	fout:write(string.format('%d\t%s\n', hashset[v], v))
+if not pretend then
+	fout = io.open(fpath, 'w')
+	for _, v in ipairs(hashset_order) do
+		fout:write(string.format('%d\t%s\n', hashset[v], v))
+	end
+	fout:close()
 end
-fout:close()
 
-if dupes == 0 then
-    print('No duplicate lines in playcount.tsv')
-else
-    print(string.format('Merged %d duplicate line%s in playcount.tsv', dupes, dupes == 1 and '' or 's'))
+if dupes ~= 0 then
+	io.stderr:write(string.format("%s %d duplicate line%s in '"..fpath:gsub(musicdir, '').."'\n",
+		pretend and 'Found' or 'Merged', dupes, dupes == 1 and '' or 's'))
 end
